pyproject.toml
[tool.poetry]
name = "wartales-trade"
version = "0.1.0"
description = ""
authors = ["Kevin Michael Engel <kevin.m.engel@gmail.com>"]
readme = "README.md"
packages = [{include = "wartales_trade"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
ttkbootstrap = "^1.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
